pyproject.toml
[project]
name = "knitwit"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2,<3",
    "flask>=3.0,<4",
    "Pillow>=12,<13",
]
